requirements.txt
google-adk
google-genai
python-dotenv
numpy
faiss-cpu
sentence-transformers
